+++ src/extension.ts
@@ -19,28 +19,30 @@ export function activate(context: ExtensionContext) {
     const showExplorerDebug: boolean = config.get(configurationSettings.showExplorerDebug);
     const autoStart: boolean = config.get(configurationSettings.autoStart);
     const autoStartDelay: number = config.get(configurationSettings.autoStartDelay);
+    const defaultSort: string = config.get(configurationSettings.defaultSort);
 
     reporter.sendTelemetryEvent('Extension Activated', {
         'showExplorer': showExplorer,
         'showExplorerDebug': showExplorerDebug,
         'autoStart': autoStart,
-        'autoStartDelay': autoStartDelay
+        'autoStartDelay': autoStartDelay,
+        'defaultSort': defaultSort
     });
 
-    setupActivityBarView(context, autoStart, autoStartDelay);
+    setupActivityBarView(context, autoStart, autoStartDelay, defaultSort);
 
     if (showExplorer) {
-        setupExplorerView(context, autoStart, autoStartDelay);
+        setupExplorerView(context, autoStart, autoStartDelay, defaultSort);
     }
 
     if (showExplorerDebug) {
-        setupDebugView(context, autoStart, autoStartDelay);
+        setupDebugView(context, autoStart, autoStartDelay, defaultSort);
     }
 }
 
-function setupDebugView(context: ExtensionContext, autoStart: boolean, autoStartDelay: number) {
+function setupDebugView(context: ExtensionContext, autoStart: boolean, autoStartDelay: number, defaultSort: string) {
     // Setup Tree Data Provider
-    const symbolsTreeDataProvider = new SymbolsTreeDataProvider(context, View.Debug, autoStart, autoStartDelay, reporter);
+    const symbolsTreeDataProvider = new SymbolsTreeDataProvider(context, View.Debug, autoStart, autoStartDelay, defaultSort, reporter);
     context.subscriptions.push(window.registerTreeDataProvider('symbolExplorerDebug', symbolsTreeDataProvider));
 
     // Setup Navigate Command
@@ -52,17 +54,21 @@ function setupDebugView(context: ExtensionContext, autoStart: boolean, autoStart
     context.subscriptions.push(refreshSymbolCommand);
 
     // Setup Sort Asc Button
-    const sortAscSymbolCommand = commands.registerCommand('symbolExplorerDebug.sortAsc', () => symbolsTreeDataProvider.refresh(Sort.Asc));
+    const sortAscSymbolCommand = commands.registerCommand('symbolExplorerDebug.sortAsc', () => symbolsTreeDataProvider.refresh(Sort.asc));
     context.subscriptions.push(sortAscSymbolCommand);
 
     // Setup Sort Desc Button
-    const sortDescSymbolCommand = commands.registerCommand('symbolExplorerDebug.sortDesc', () => symbolsTreeDataProvider.refresh(Sort.Desc));
+    const sortDescSymbolCommand = commands.registerCommand('symbolExplorerDebug.sortDesc', () => symbolsTreeDataProvider.refresh(Sort.desc));
     context.subscriptions.push(sortDescSymbolCommand);
+
+    // Setup Collapse Button
+    // const collapseSymbolCommand = commands.registerCommand('symbolExplorerDebug.collapse', () => symbolsTreeDataProvider.refresh(Sort[defaultSort], true));
+    // context.subscriptions.push(collapseSymbolCommand);
 }
 
-function setupExplorerView(context: ExtensionContext, autoStart: boolean, autoStartDelay: number) {
+function setupExplorerView(context: ExtensionContext, autoStart: boolean, autoStartDelay: number, defaultSort: string) {
     // Setup Tree Data Provider
-    const symbolsTreeDataProvider = new SymbolsTreeDataProvider(context, View.Explorer, autoStart, autoStartDelay, reporter);
+    const symbolsTreeDataProvider = new SymbolsTreeDataProvider(context, View.Explorer, autoStart, autoStartDelay, defaultSort, reporter);
     context.subscriptions.push(window.registerTreeDataProvider('symbolExplorer', symbolsTreeDataProvider));
 
     // Setup Navigate Command
@@ -74,17 +80,21 @@ function setupExplorerView(context: ExtensionContext, autoStart: boolean, autoSt
     context.subscriptions.push(refreshSymbolCommand);
 
     // Setup Sort Asc Button
-    const sortAscSymbolCommand = commands.registerCommand('symbolExplorer.sortAsc', () => symbolsTreeDataProvider.refresh(Sort.Asc));
+    const sortAscSymbolCommand = commands.registerCommand('symbolExplorer.sortAsc', () => symbolsTreeDataProvider.refresh(Sort.asc));
     context.subscriptions.push(sortAscSymbolCommand);
 
     // Setup Sort Desc Button
-    const sortDescSymbolCommand = commands.registerCommand('symbolExplorer.sortDesc', () => symbolsTreeDataProvider.refresh(Sort.Desc));
+    const sortDescSymbolCommand = commands.registerCommand('symbolExplorer.sortDesc', () => symbolsTreeDataProvider.refresh(Sort.desc));
     context.subscriptions.push(sortDescSymbolCommand);
+
+    // Setup Collapse Button
+    // const collapseSymbolCommand = commands.registerCommand('symbolExplorer.collapse', () => symbolsTreeDataProvider.refresh(Sort[defaultSort], true));
+    // context.subscriptions.push(collapseSymbolCommand);
 }
 
-function setupActivityBarView(context: ExtensionContext, autoStart: boolean, autoStartDelay: number) {
+function setupActivityBarView(context: ExtensionContext, autoStart: boolean, autoStartDelay: number, defaultSort: string) {
     // Setup Tree Data Provider
-    const symbolsTreeDataProvider = new SymbolsTreeDataProvider(context, View.View, autoStart, autoStartDelay, reporter);
+    const symbolsTreeDataProvider = new SymbolsTreeDataProvider(context, View.View, autoStart, autoStartDelay, defaultSort, reporter);
     context.subscriptions.push(window.registerTreeDataProvider('symbolExplorerView', symbolsTreeDataProvider));
 
     // Setup Navigate Command
@@ -96,12 +106,16 @@ function setupActivityBarView(context: ExtensionContext, autoStart: boolean, aut
     context.subscriptions.push(refreshSymbolCommand);
 
     // Setup Sort Asc Button
-    const sortAscSymbolCommand = commands.registerCommand('symbolExplorerView.sortAsc', () => symbolsTreeDataProvider.refresh(Sort.Asc));
+    const sortAscSymbolCommand = commands.registerCommand('symbolExplorerView.sortAsc', () => symbolsTreeDataProvider.refresh(Sort.asc));
     context.subscriptions.push(sortAscSymbolCommand);
 
     // Setup Sort Desc Button
-    const sortDescSymbolCommand = commands.registerCommand('symbolExplorerView.sortDesc', () => symbolsTreeDataProvider.refresh(Sort.Desc));
+    const sortDescSymbolCommand = commands.registerCommand('symbolExplorerView.sortDesc', () => symbolsTreeDataProvider.refresh(Sort.desc));
     context.subscriptions.push(sortDescSymbolCommand);
+
+    // Setup Collapse Button
+    // const collapseSymbolCommand = commands.registerCommand('symbolExplorerView.collapse', () => symbolsTreeDataProvider.refresh(Sort[defaultSort], true));
+    // context.subscriptions.push(collapseSymbolCommand);
 }
 
 // this method is called when your extension is deactivated